Django/requirements.txt
Django==2.2.1
django-cors-headers==3.0.1
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
Markdown==3.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
